+++ CHANGELOG.md
@@ -30,11 +30,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed backup restore not working if one of the backup-parts already existed on the cluster
 - Fixed K8s migrations to use correct generated DB table instances
 - Deleted invalid SpaceHistory entries from K8s.
-- Fixed left-over tiebreaker after evacuation of sinlge DRBD resource
+- Fixed left-over tiebreaker after evacuation of single DRBD resource
 - Fixed negative result of an allocation granularity calculation
 - Fixed infinite loop and limited range of power-of-2 calculations
 - Workaround for connections where the attached peer object is missing
 - Only show warnings of missing cgroup (v1) if the user actually tries to use it
+- Fixed undelete resources (also remove volume DELETE and DRBD_DELETE flags)
 
 ## [1.26.2] - 2024-02-28